# Improve span creation performance by 10-20%