It's cheaper to generate random u128 and u64 than it is to generate [u8; 16] and [u8; 8].

Changing `TraceId::From` and `SpanId::From` to be from these primitive types, rather than arrays, improves `start-end-span` benchmark performance by 10% in the sampling case and 21% in the non-sampling case.

```text
start-end-span/always-sample
                        time:   [517.71 ns 535.68 ns 551.39 ns]
                        change: [-14.991% -10.282% -5.3887%] (p = 0.00 < 0.05)
                        Performance has improved.
start-end-span/never-sample
                        time:   [133.18 ns 133.46 ns 133.76 ns]
                        change: [-22.160% -21.965% -21.769%] (p = 0.00 < 0.05)
                        Performance has improved.
```

## Changes

- Change TraceId::From and SpanId::From to be from primitive types, rather than arrays with an assumed byte ordering.
- DatadogPropagator had to change too, but it will be faster as it's no longer converting to big-endian bytes just to call TraceId/SpanId::From.

## Merge requirement checklist

* [x] [CONTRIBUTING](https://github.com/open-telemetry/opentelemetry-rust/blob/main/CONTRIBUTING.md) guidelines followed
* [x] Unit tests added/updated (if applicable)
* [ ] Appropriate `CHANGELOG.md` files updated for non-trivial, user-facing changes
* [ ] Changes in public API reviewed (if applicable)